# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Benefit:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scheme_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ChildCriteria:
    properties:
      school_level:
        type: string
    type: object
  models.Criteria:
    properties:
      employment_status:
        type: string
      has_children:
        $ref: '#/definitions/models.ChildCriteria'
      marital_status:
        type: string
    type: object
  models.EligibleSchemesResponse:
    properties:
      applicant_id:
        type: string
      schemes:
        items:
          $ref: '#/definitions/models.SchemeResponse'
        type: array
    type: object
...
  models.Scheme:
    properties:
      benefits:
        items:
          $ref: '#/definitions/models.Benefit'
        type: array
      created_at:
        type: string
      criteria:
        $ref: '#/definitions/models.Criteria'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.SchemeResponse:
    properties:
      benefits:
        items:
          $ref: '#/definitions/models.Benefit'
        type: array
      created_at:
        type: string
      criteria:
        $ref: '#/definitions/models.Criteria'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
paths:
...
  /api/schemes:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all financial assistance schemes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SchemeResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all schemes
      tags:
      - schemes
    post:
      consumes:
      - application/json
      description: Add a new financial assistance scheme
      parameters:
      - description: Scheme information
        in: body
        name: scheme
        required: true
        schema:
          $ref: '#/definitions/models.Scheme'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SchemeResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new scheme
      tags:
      - schemes
  /api/schemes/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a scheme from the system
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Scheme not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete scheme
      tags:
      - schemes
    get:
      consumes:
      - application/json
      description: Retrieve a specific scheme by its ID
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SchemeResponse'
        "404":
          description: Scheme not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get scheme by ID
      tags:
      - schemes
    put:
      consumes:
      - application/json
      description: Update an existing scheme's information
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated scheme information
        in: body
        name: scheme
        required: true
        schema:
          $ref: '#/definitions/models.Scheme'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SchemeResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Scheme not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update scheme
      tags:
      - schemes
  /api/schemes/eligible:
    get:
      consumes:
      - application/json
      description: Retrieve all schemes that an applicant is eligible for
      parameters:
      - description: Applicant ID
        in: query
        name: applicant
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EligibleSchemesResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Applicant not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get eligible schemes for an applicant
      tags:
      - schemes
schemes:
- http
swagger: "2.0"